+++ CHANGELOG.md
@@ -192,6 +192,7 @@
 ### Query-tee
 
 * [CHANGE] If you configure multiple secondary backends and enable comparisons, query-tee reports comparison results of the preferred backend against each of the secondaries. #13022
+* [CHANGE] Add backend configuration options for request proportion sampling and time-based query filtering. #13037
 
 ## 2.17.1